Task#23

+++ src/app/shared/services/produto.service.ts
@@ -1,15 +1,16 @@
-import { switchMap } from 'rxjs/operators';
+import { catchError, switchMap } from 'rxjs/operators';
 import { Produto } from './../model/produto';
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { Observable } from 'rxjs';
+import { Observable, throwError } from 'rxjs';
+import { environment } from 'src/environments/environment';
 
 @Injectable({
   providedIn: 'root'
 })
 export class ProdutoService {
 
-  private URL_PRODUTOS = 'http://localhost:3000/produtos';
+  private URL_PRODUTOS = environment.URL_API + '/produtos';
 
   constructor(private http: HttpClient) { }
 
@@ -22,16 +23,43 @@ export class ProdutoService {
   }
 
   localizar(id: number): Observable<number> {
-    return this.http.get<number>(`${this.URL_PRODUTOS}/produtos/${id}`);
+    return this.http.get<number>(`${this.URL_PRODUTOS}/${id}`);
   }
 
   remover(id: number): Observable<void> {
-    return this.http.delete<void>(`${this.URL_PRODUTOS}/${id}`);
+    const confirmation = confirm('Tem certeza de que deseja excluir este produto?');
+    if (!confirmation) {
+      return new Observable(); // Ou qualquer Observable vazio
+    }
+
+    return this.http.delete<void>(`${this.URL_PRODUTOS}/${id}`)
+      .pipe(
+        catchError(this.handleError)
+      );
+  }
+
+  private handleError(error: any) {
+    console.error('Erro na requisição:', error);
+    const errorMessage = 'Erro na requisição. Por favor, tente novamente mais tarde.';
+    return throwError(() => new Error(errorMessage));
   }
 
   editar(idproduto: number, produto: Produto): Observable<Produto> {
     return this.http.put<Produto>(`${this.URL_PRODUTOS}/${idproduto}`, produto);
   }
+
+  pesquisarPorNome(nome: string): Observable<Produto[]> {
+    return this.http.get<Produto[]>(`${this.URL_PRODUTOS}?nome=${nome}`);
+  }
+
+  pesquisarPorCategoria(categoria: string): Observable<Produto[]> {
+    return this.http.get<Produto[]>(`${this.URL_PRODUTOS}?categoria=${categoria}`);
+  }
+
+  // existenciaProduto(produtoNome: string): Observable<boolean> {
+  //   const url = `${this.URL_PRODUTOS}/exists/${produtoNome}`;
+  //   return this.http.get<boolean>(url);
+  // }
 }
